requirements.txt
boto
boto3
botocore
PyVmomi
pynetbox
pyyaml
ansible-vault
